Move fastfetch-specific files
