+++ GVFS/FastFetch/FastFetch.csproj
@@ -60,8 +60,10 @@
     <Compile Include="CheckoutPrefetcher.cs" />
     <Compile Include="FastFetchVerb.cs" />
     <Compile Include="..\GVFS.PlatformLoader\PlatformLoader.Windows.cs" />
+    <Compile Include="Index.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="WorkingTree.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
