+++ GVFS/FastFetch/WorkingTree.cs
@@ -1,12 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using GVFS.Common;
 
-namespace GVFS.Common.Prefetch
+namespace FastFetch
 {
     public static class WorkingTree
     {
